requirements.txt
python-dotenv
paho_mqtt
pyserial
smbus2